move GrpcQueryHandler into distributed.rs
