--- src/frontend/src/instance/distributed/grpc.rs
@@ -1,65 +1,0 @@
-// Copyright 2023 Greptime Team
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use api::v1::ddl_request::Expr as DdlExpr;
-use api::v1::greptime_request::Request;
-use async_trait::async_trait;
-use common_query::Output;
-use meta_client::rpc::TableName;
-use servers::query_handler::grpc::GrpcQueryHandler;
-use session::context::QueryContextRef;
-use snafu::OptionExt;
-
-use crate::error::{self, Result};
-use crate::instance::distributed::DistInstance;
-
-#[async_trait]
-impl GrpcQueryHandler for DistInstance {
-    type Error = error::Error;
-
-    async fn do_query(&self, request: Request, ctx: QueryContextRef) -> Result<Output> {
-        match request {
-            Request::Insert(request) => self.handle_dist_insert(request, ctx).await,
-            Request::Delete(request) => self.handle_dist_delete(request, ctx).await,
-            Request::Query(_) => {
-                unreachable!("Query should have been handled directly in Frontend Instance!")
-            }
-            Request::Ddl(request) => {
-                let expr = request.expr.context(error::IncompleteGrpcResultSnafu {
-                    err_msg: "Missing 'expr' in DDL request",
-                })?;
-                match expr {
-                    DdlExpr::CreateDatabase(expr) => self.handle_create_database(expr, ctx).await,
-                    DdlExpr::CreateTable(mut expr) => {
-                        // TODO(LFC): Support creating distributed table through GRPC interface.
-                        // Currently only SQL supports it; how to design the fields in CreateTableExpr?
-                        let _ = self.create_table(&mut expr, None).await;
-                        Ok(Output::AffectedRows(0))
-                    }
-                    DdlExpr::Alter(expr) => self.handle_alter_table(expr).await,
-                    DdlExpr::DropTable(expr) => {
-                        let table_name =
-                            TableName::new(&expr.catalog_name, &expr.schema_name, &expr.table_name);
-                        self.drop_table(table_name).await
-                    }
-                    DdlExpr::FlushTable(expr) => {
-                        let table_name =
-                            TableName::new(&expr.catalog_name, &expr.schema_name, &expr.table_name);
-                        self.flush_table(table_name, expr.region_id).await
-                    }
-                }
-            }
-        }
-    }
-}
